+++ nonconform/estimation/extreme_conformal.py
@@ -196,7 +196,7 @@ class ExtremeConformalDetector(BaseConformalDetector):
                 self.detector_set,
                 total=len(self.detector_set),
                 desc=f"Aggregating {len(self.detector_set)} models",
-                disable=not logger.isEnabledFor(logging.INFO),
+                disable=not logger.isEnabledFor(logging.DEBUG),
             )
         ]