+++ nonconform/estimation/weighted_conformal.py
@@ -169,7 +169,7 @@ class WeightedConformalDetector(BaseConformalDetector):
                 self.detector_set,
                 total=len(self.detector_set),
                 desc=f"Aggregating {len(self.detector_set)} models",
-                disable=not logger.isEnabledFor(logging.INFO),
+                disable=not logger.isEnabledFor(logging.DEBUG),
             )
         ]